Refactor Stimulus Controllers with regard to utilities.js
