+++ app/javascript/application.js
@@ -10,4 +10,4 @@ import "Sortable"
 import "utilities"
 import "flatpickr"
 import "confirmDate"
-import "expts/hotwire-events"
+import "hotwire-events"
