+++ app/javascript/controllers/task_controller.js
@@ -1,8 +1,8 @@
 import AbstractTask from "controllers/abstract_task_controller"
 
-import logHotwireEvents from "expts/hotwire-events"
+import logHotwireEvents from "hotwire-events"
 
-import { log, creEl, qs } from "utilities"
+import { log, creEl, qs, hasCls, replCls, iae } from "utilities"
 
 export default class extends AbstractTask {
   static targets = ["subtasksBtn", "subtasksBtnIcon", "subtasksContainer"]
@@ -35,7 +35,7 @@ export default class extends AbstractTask {
   onHandleGrabbed(e) {
     // Do not cancel the event: e.preventDefault()
     super.onHandleGrabbed(e)
-    if (this.subtasksContainerTarget.classList.contains("show")) {
+    if (hasCls(this.subtasksContainerTarget, "show")) {
       // Then close it
       this.subtasksBtnTarget.click()
     }
@@ -43,7 +43,7 @@ export default class extends AbstractTask {
 
   doSubtasksOpenForNewSubtask(e) {
     // e.preventDefault()
-    if (!this.subtasksContainerTarget.classList.contains("show")) {
+    if (!hasCls(this.subtasksContainerTarget, "show")) {
       // Then open it
       // const event = new Event("click")
       // this.subtasksBtnTarget.dispatchEvent(event)
@@ -53,18 +53,12 @@ export default class extends AbstractTask {
 
   onSubtasksOpen(e) {
     e.preventDefault()
-    this.subtasksBtnIconTarget.classList.replace(
-      "fa-angles-down",
-      "fa-angles-up"
-    )
+    replCls(this.subtasksBtnIconTarget, "fa-angles-down", "fa-angles-up")
   }
 
   onSubtasksClose(e) {
     e.preventDefault()
-    this.subtasksBtnIconTarget.classList.replace(
-      "fa-angles-up",
-      "fa-angles-down"
-    )
+    replCls(this.subtasksBtnIconTarget, "fa-angles-up", "fa-angles-down")
   }
 
   onCheckboxClicked(e) {
@@ -75,6 +69,7 @@ export default class extends AbstractTask {
       this.subtasks.length &&
       this.allSubtasksChecked()
     ) {
+      // Would not '-1' sufice?
       this.subtasks[this.subtasks.length - 1].uncheckIt()
     } else if (this.isChecked() && this.allSubtasksNotChecked()) {
       this.subtasks.forEach((subtask) => {
@@ -134,7 +129,7 @@ export default class extends AbstractTask {
       })
       .then((data) => {
         this.hiddenFormCont.innerHTML = data.form
-        this.form = this.hiddenFormCont.getElementsByTagName("form")[0]
+        this.form = qs("form", this.hiddenFormCont)
 
         const formData = new FormData(this.form)
 
@@ -225,6 +220,6 @@ export default class extends AbstractTask {
       id: `${this.element.id}_update_form_container`,
       class: "d-none",
     })
-    qs("body").insertAdjacentElement("beforeend", this.hiddenFormCont)
+    iae("beforeend", this.hiddenFormCont, qs("body"))
   }
 }
